requirements.txt
# conda create --name panda_rl_env --file requirements.txt --channel conda-forge

python==3.7.11
numpy==1.21.2
pytorch 
torchvision 
matplotlib==3.5
matplotlib-base==3.5.1
pillow==9.0.1
ipython==7.31.1
plotly==2.7
gym==0.19.0
stable-baselines3==1.4.0
optuna==2.10

pybullet==3.2.1
torchaudio
panda-gym==2.0.1
opencv-python==3.4.17.63
